Extract CoreWebViewThemeSwitcher
